@@ -31,141 +29,151 @@
     "npm": ">= 3.0.0"
   },
   "dependencies": {
-    "bcrypt-nodejs": "0.0.3",
-    "bluebird": "^3.5.0",
-    "body-parser": "~1.17.1",
-    "config": "^1.26.1",
-    "cookie": "^0.3.1",
-    "cookie-parser": "^1.4.3",
-    "debug": "~2.6.6",
-    "errorhandler": "^1.5.0",
-    "express": "~4.15.2",
-    "express-promise-router": "^1.1.1",
-    "graphql": "^0.9.6",
-    "graphql-server-express": "^0.7.2",
-    "graphql-subscriptions": "^0.3.1",
-    "graphql-tools": "^0.11.0",
-    "i18next": "^8.1.0",
-    "i18next-express-middleware": "^1.0.5",
-    "i18next-node-fs-backend": "^1.0.0",
-    "jsonwebtoken": "^7.4.0",
-    "lodash": "^4.17.4",
-    "method-override": "^2.3.8",
-    "moment": "^2.18.1",
-    "mongodb": "~2.2.26",
-    "morgan": "^1.8.1",
-    "morgan-debug": "^2.0.0",
-    "passport": "^0.3.2",
-    "passport-anonymous": "^1.0.1",
-    "passport-jwt": "^2.2.1",
-    "permission": "^1.0.4",
-    "serve-favicon": "~2.4.2",
-    "stream-to-promise": "^2.2.0",
-    "subscriptions-transport-ws": "^0.6.0"
+    "@std/esm": "^0.8.3",
+    "bcrypt-nodejs": "~0.0.3",
+    "bluebird": "~3.5.0",
+    "body-parser": "~1.18.0",
+    "config": "~1.26.2",
+    "cookie": "~0.3.1",
+    "cookie-parser": "~1.4.3",
+    "debug": "~3.0.1",
+    "errorhandler": "~1.5.0",
+    "express": "~4.15.4",
+    "express-promise-router": "~2.0.0",
+    "graphql": "~0.11.3",
+    "graphql-server-express": "~1.1.2",
+    "graphql-subscriptions": "~0.4.4",
+    "graphql-tools": "~1.2.2",
+    "i18next": "~9.0.0",
+    "i18next-express-middleware": "~1.0.5",
+    "i18next-node-fs-backend": "~1.0.0",
+    "jsonwebtoken": "~8.0.0",
+    "lodash": "~4.17.4",
+    "method-override": "~2.3.9",
+    "moment": "~2.18.1",
+    "mongodb": "~2.2.31",
+    "morgan": "~1.8.2",
+    "morgan-debug": "~2.0.0",
+    "passport": "~0.4.0",
+    "passport-anonymous": "~1.0.1",
+    "passport-jwt": "~3.0.0",
+    "permission": "~1.0.4",
+    "serve-favicon": "~2.4.3",
+    "stream-to-promise": "~2.2.0",
+    "subscriptions-transport-ws": "~0.8.3"
   },
   "devDependencies": {
-    "apidoc": "^0.17.5",
-    "apollo-client": "^1.2.2",
-    "autoprefixer": "^7.0.1",
-    "babel-cli": "^6.24.1",
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "7.0.0",
-    "babel-plugin-istanbul": "^4.1.1",
-    "babel-plugin-lodash": "^3.2.11",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.4.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2015-loose": "^8.0.0",
-    "babel-preset-stage-1": "^6.24.1",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.22.0",
-    "babel-runtime": "^6.23.0",
-    "chai": "^3.5.0",
-    "chalk": "^1.1.3",
-    "chromedriver": "^2.27.2",
-    "clipboard-copy": "^1.2.0",
-    "connect-flash": "^0.1.1",
-    "connect-history-api-fallback": "^1.3.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^4.0.0",
-    "cross-spawn": "^5.0.1",
-    "crypto-browserify": "^3.11.0",
-    "css-loader": "^0.28.1",
-    "download-as-file": "^1.1.0",
-    "ejs-loader": "^0.3.0",
-    "eslint": "^3.19.0",
-    "eslint-config-standard": "^6.2.1",
-    "eslint-friendly-formatter": "^2.0.7",
-    "eslint-loader": "^1.7.1",
-    "eslint-plugin-html": "^2.0.0",
-    "eslint-plugin-nodeca": "^1.0.3",
-    "eslint-plugin-promise": "^3.4.0",
-    "eslint-plugin-standard": "^2.0.1",
-    "eventsource-polyfill": "^0.9.6",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.11.1",
-    "friendly-errors-webpack-plugin": "^1.1.3",
-    "html-loader": "^0.4.5",
-    "html-webpack-plugin": "^2.28.0",
-    "http-proxy-middleware": "^0.17.3",
-    "i18next-resource-store-loader": "0.1.1",
-    "img-loader": "^2.0.0",
-    "inject-loader": "^3.0.0",
-    "jquery": "^3.2.1",
-    "json-loader": "^0.5.4",
-    "json2csv": "^3.7.3",
-    "karma": "^1.4.1",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.3.0",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-phantomjs-shim": "^1.4.0",
-    "karma-sinon-chai": "^1.3.1",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.30",
-    "karma-webpack": "^2.0.2",
-    "lodash-webpack-plugin": "^0.11.3",
-    "lolex": "^1.5.2",
-    "mocha": "^3.2.0",
-    "nightwatch": "^0.9.12",
-    "node-sass": "^4.5.2",
-    "optimize-css-assets-webpack-plugin": "^2.0.0",
-    "ora": "^1.3.0",
-    "phantomjs-prebuilt": "^2.1.14",
-    "postcss-loader": "^2.0.1",
-    "pug": "^2.0.0-rc.1",
-    "pug-loader": "^2.3.0",
-    "rimraf": "^2.6.1",
-    "sass-loader": "^6.0.3",
-    "selenium-server": "^3.0.1",
-    "semver": "^5.3.0",
-    "shelljs": "^0.7.6",
-    "sinon": "^2.1.0",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.17.0",
-    "stylus": "^0.54.5",
-    "stylus-loader": "^3.0.1",
-    "url-loader": "^0.5.8",
-    "vee-validate": "^2.0.0-rc.3",
-    "vue": "^2.3.3",
-    "vue-apollo": "^2.1.0-beta.8",
-    "vue-loader": "^12.0.4",
-    "vue-router": "^2.5.3",
-    "vue-style-loader": "^3.0.1",
-    "vue-template-compiler": "^2.3.3",
-    "vuetify": "^0.12.0",
-    "webpack": "^2.5.1",
-    "webpack-bundle-analyzer": "^2.2.1",
-    "webpack-dev-middleware": "^1.10.0",
-    "webpack-dev-server": "^2.4.5",
-    "webpack-hot-middleware": "^2.18.0",
-    "webpack-merge": "^4.1.0",
-    "yaml-loader": "^0.4.0"
+    "apidoc": "~0.17.6",
+    "apollo-client": "~1.9.2",
+    "autoprefixer": "~7.1.4",
+    "babel-cli": "~6.26.0",
+    "babel-core": "~6.26.0",
+    "babel-eslint": "~7.2.3",
+    "babel-loader": "7.1.2",
+    "babel-plugin-istanbul": "~4.1.4",
+    "babel-plugin-lodash": "~3.2.11",
+    "babel-plugin-transform-decorators-legacy": "~1.3.4",
+    "babel-plugin-transform-runtime": "~6.23.0",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "~1.6.0",
+    "babel-preset-es2015": "~6.24.1",
+    "babel-preset-es2015-loose": "~8.0.0",
+    "babel-preset-stage-1": "~6.24.1",
+    "babel-preset-stage-2": "~6.24.1",
+    "babel-register": "^6.26.0",
+    "babel-runtime": "~6.26.0",
+    "chai": "~4.1.2",
+    "chalk": "~2.1.0",
+    "chromedriver": "~2.32.2",
+    "clipboard-copy": "~1.2.0",
+    "connect-flash": "~0.1.1",
+    "connect-history-api-fallback": "~1.3.0",
+    "copy-webpack-plugin": "~4.0.1",
+    "cross-env": "~5.0.5",
+    "cross-spawn": "~5.1.0",
+    "crypto-browserify": "~3.11.1",
+    "css-loader": "~0.28.7",
+    "download-as-file": "~1.1.0",
+    "ejs-loader": "~0.3.0",
+    "es6-promisify": "^5.0.0",
+    "eslint": "~4.6.1",
+    "eslint-config-standard": "~10.2.1",
+    "eslint-friendly-formatter": "~3.0.0",
+    "eslint-loader": "~1.9.0",
+    "eslint-plugin-html": "~3.2.1",
+    "eslint-plugin-import": "~2.7.0",
+    "eslint-plugin-node": "~5.1.1",
+    "eslint-plugin-nodeca": "~1.0.3",
+    "eslint-plugin-promise": "~3.5.0",
+    "eslint-plugin-standard": "~3.0.1",
+    "eventsource-polyfill": "~0.9.6",
+    "extract-text-webpack-plugin": "~3.0.0",
+    "file-loader": "~0.11.2",
+    "friendly-errors-webpack-plugin": "~1.6.1",
+    "geckodriver": "^1.8.1",
+    "html-loader": "~0.5.1",
+    "html-webpack-plugin": "~2.30.1",
+    "http-proxy-middleware": "~0.17.4",
+    "i18next-resource-store-loader": "0.1.2",
+    "img-loader": "~2.0.0",
+    "inject-loader": "^3.0.1",
+    "jquery": "~3.2.1",
+    "json-loader": "~0.5.7",
+    "json2csv": "~3.11.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage": "~1.1.1",
+    "karma-mocha": "~1.3.0",
+    "karma-phantomjs-shim": "~1.4.0",
+    "karma-sinon-chai": "~1.3.2",
+    "karma-sourcemap-loader": "~0.3.7",
+    "karma-spec-reporter": "0.0.31",
+    "karma-webpack": "~2.0.4",
+    "lodash-webpack-plugin": "~0.11.4",
+    "lolex": "~2.1.2",
+    "mocha": "~3.5.2",
+    "nightwatch": "~0.9.16",
+    "node-sass": "~4.5.3",
+    "optimize-css-assets-webpack-plugin": "~3.1.1",
+    "ora": "~1.3.0",
+    "phantomjs-prebuilt": "~2.1.15",
+    "postcss-loader": "~2.0.6",
+    "pug": "~2.0.0-rc.4",
+    "pug-loader": "~2.3.0",
+    "rimraf": "~2.6.1",
+    "sass-loader": "~6.0.6",
+    "selenium-server": "~3.5.3",
+    "semver": "~5.4.1",
+    "shelljs": "~0.7.8",
+    "sinon": "~3.2.1",
+    "sinon-chai": "~2.13.0",
+    "style-loader": "~0.18.2",
+    "stylus": "~0.54.5",
+    "stylus-loader": "~3.0.1",
+    "url-loader": "~0.5.9",
+    "vee-validate": "~2.0.0-rc.14",
+    "vue": "~2.4.2",
+    "vue-apollo": "~2.1.0-rc.4",
+    "vue-loader": "~13.0.4",
+    "vue-router": "~2.7.0",
+    "vue-style-loader": "~3.0.1",
+    "vue-template-compiler": "~2.4.2",
+    "vue-test-utils": "git+https://github.com/vuejs/vue-test-utils.git",
+    "vuetify": "~0.15",
+    "webpack": "~3.5.6",
+    "webpack-bundle-analyzer": "~2.9.0",
+    "webpack-dev-middleware": "~1.12.0",
+    "webpack-dev-server": "~2.7.1",
+    "webpack-hot-middleware": "~2.19.1",
+    "webpack-merge": "~4.1.0",
+    "yaml-loader": "~0.5.0"
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
     "not ie <= 8"
-  ]
+  ],
+  "@std/esm": {
+    "cjs": true
+  }
 }
