@@ -1,59 +1,58 @@
 <template lang="pug">
-v-card
-	v-card-row.darken-1(v-bind:class="cardType.color")
-		v-card-title
-			span.white--text {{ trans(cardType.label) }}
-			v-spacer
+v-card.detail-card
+	v-toolbar.white--text.darken-1(v-bind:class="cardType.color")
+		v-toolbar-title#title-label {{ trans(cardType.label) }}
+
 	v-card-text
 		v-layout(row,wrap)
 			v-flex(xs12)
-				v-text-field(
+				v-text-field#label-input(
 					:label="trans('items:item.form.label.field')"
 					:data-vv-as="trans('items:item.form.label.field')",
 					v-validate="'required'",
 					name="label",
-					v-model="line.label",
+					v-model="lineToModify.label",
 					v-bind:rules="labelValidation")
 
 			v-flex(xs12)
-				v-select(
+				v-select#group-select(
 					:label="trans('items:item.form.group.field')",
-          v-bind:items="selectGroups",
-	        v-model="line.group",
+					v-bind:items="selectGroups",
+					v-model="lineToModify.group",
 					item-text="name"
 					item-value="value")
 					template(slot="item",scope="data")
 						template(v-if="data.item.value") {{ data.item.name }}
 						div.groupSelect(v-if="!data.item.value")
-							span {{ data.item.name }}
+							span#empty-item {{ data.item.name }}
 
-			v-flex(xs12,v-if="line.group === ''")
-				v-text-field(
-	        :label="trans('items:item.form.group.new')"
+			v-flex(xs12,v-if="lineToModify.group === ''")
+				v-text-field#group-input(
+					:label="trans('items:item.form.group.new')"
 					v-model="newGroup")
 
-			template(v-if="line.type == 'text'")
+			template(v-if="lineToModify.type == 'text'")
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#text-input(
 						:label="trans('items:item.form.text.field')",
 						v-model="clearInformation.text",
 						multi-line,
 						auto-grow)
 
-			template(v-if="line.type == 'card'")
+			template(v-if="lineToModify.type == 'card'")
 				v-flex(xs12)
-					v-select(
+					v-select#type-of-card-select(
 						:label="trans('items:item.form.type.field')",
 						v-bind:items="typeOfCard",
 						v-model="clearInformation.type")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#name-on-card-input(
 						:label="trans('items:item.form.nameOnCard.field')",
 						v-model="clearInformation.nameOnCard")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#card-number-input(
 						:label="trans('items:item.form.cardNumber.field')",
 						:append-icon="cardNumberVisibility ? 'visibility' : 'visibility_off'"
 						:append-icon-cb="() => (cardNumberVisibility = !cardNumberVisibility)"
@@ -61,7 +60,7 @@ v-card
 						v-model="clearInformation.cardNumber")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#cvv-input(
 						:label="trans('items:item.form.cvv.field')",
 						:append-icon="cvvVisibility ? 'visibility' : 'visibility_off'"
 						:append-icon-cb="() => (cvvVisibility = !cvvVisibility)"
@@ -69,26 +68,26 @@ v-card
 						v-model="clearInformation.cvv")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#expiry-input(
 						:label="trans('items:item.form.expiry.field')",
 						v-model="clearInformation.expiry")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#code-input(
 						:label="trans('items:item.form.code.field')",
 						:append-icon="codeVisibility ? 'visibility' : 'visibility_off'"
 						:append-icon-cb="() => (codeVisibility = !codeVisibility)"
 						:type="codeVisibility ? 'text' : 'password'"
 						v-model="clearInformation.code")
 
-			template(v-if="line.type == 'password'")
+			template(v-if="lineToModify.type == 'password'")
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#username-input(
 						:label="trans('items:item.form.username.field')",
 						v-model="clearInformation.username")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#password-input(
 						:label="trans('items:item.form.password.field')",
 						:append-icon="passwordVisibility ? 'visibility' : 'visibility_off'"
 						:append-icon-cb="() => (passwordVisibility = !passwordVisibility)"
@@ -96,12 +95,12 @@ v-card
 						v-model="clearInformation.password")
 
 				v-flex(xs12)
-					v-text-field(
+					v-text-field#siteurl-input(
 						:label="trans('items:item.form.siteUrl.field')",
 						v-model="clearInformation.siteUrl")
 
 			v-flex(xs12)
-				v-text-field(
+				v-text-field#notes-input(
 					:label="trans('items:item.form.notes.field')",
 					v-model="clearInformation.notes",
 					multi-line,
@@ -109,14 +108,14 @@ v-card
 
 			v-flex(xs12)
 				.text-xs-center
-					v-btn(primary,light,v-on:click.native="submitForm()") {{ trans('items:item.form.button.field') }}
-					template(v-if="line.type == 'password'")
+					v-btn#detail-button(primary,dark,v-on:click.native="submitForm()") {{ trans('items:item.form.button.field') }}
+					template(v-if="lineToModify.type == 'password'")
 						v-spacer
-						v-btn(primary,light,v-on:click.native="generatePassword()") {{ trans('items:item.form.button.generate') }}
+						v-btn#generate-password(primary,dark,v-on:click.native="generatePassword()") {{ trans('items:item.form.button.generate') }}
 </template>
 
 <script type="text/babel">
-import {pick, map} from 'lodash'
+import {pick, map, cloneDeep} from 'lodash'
 import {cardTypeMapping, updateLine, decryptLine, encryptLine, generate} from './ItemService'
 import getGroups from './getGroups.gql'
 
@@ -125,6 +124,7 @@ export default {
   name: 'item-detail',
   data () {
     return {
+      lineToModify: cloneDeep(this.line),
       cardNumberVisibility: false,
       cvvVisibility: false,
       codeVisibility: false,
@@ -140,24 +140,30 @@ export default {
     groups: getGroups
   },
   methods: {
-    submitForm () {
+    async submitForm () {
       if (this.veeErrors.any()) {
         return
       }
 
-      const line = pick(this.line, ['_id', 'type', 'label', 'group', '_rev'])
+      const line = pick(this.lineToModify, ['_id', 'type', 'label', 'group', '_rev'])
 
       if (this.newGroup) {
         line.group = this.newGroup
       }
 
-      encryptLine(this.clearInformation)
-      .then(encryptedInformation => (line.encryption = encryptedInformation))
-      .then(() => updateLine(this, line))
-      .then(() => this.$router.push('/items'))
+      line.encryption = await encryptLine(this.clearInformation)
+
+      await updateLine(this, line)
+      this.$router.push('/items')
     },
-    generatePassword () {
-      generate().then(password => (this.clearInformation.password = password))
+    async generatePassword () {
+      this.clearInformation.password = await generate()
+    },
+    async decryptClearInformation (val) {
+      if (!val.type) {
+        val = {type: 'text'}
+      }
+      this.clearInformation = await decryptLine(val)
     }
   },
   computed: {
@@ -174,18 +180,13 @@ export default {
       return result
     },
     cardType () {
-      return cardTypeMapping[this.line.type || 'text']
+      return cardTypeMapping[this.lineToModify.type || 'text']
     }
   },
   watch: {
-    line: {
+    lineToModify: {
       immediate: true,
-      handler: function (val) {
-        if (!val.type) {
-          val = {type: 'text'}
-        }
-        decryptLine(val).then(newValue => (this.clearInformation = newValue))
-      }
+      handler: async val => this.decryptClearInformation(val)
     }
   }
 }
