@@ -7,7 +7,7 @@
 		h3 {{ trans('about.licence') }}
 
 		p MIT License
-		p Copyright (c) 2016 Ulrich Van Den Hekke
+		p Copyright (c) 2017 Ulrich Van Den Hekke
 
 		p
 		| Permission is hereby granted, free of charge, to any person obtaining a copy
