@@ -19,10 +19,10 @@
 						v-model="password"
 						v-bind:rules="passwordValidation")
 					.text-xs-center
-						v-btn(primary,light,v-on:click.native="submitForm()") {{ trans('user:login.dialog.button.connect') }}
+						v-btn#login-button(primary,dark,v-on:click.native="submitForm()") {{ trans('user:login.dialog.button.connect') }}
 					div.mt-5
 						| 	{{ trans('user:login.dialog.button.noaccount') }}
-						router-link(to="/register") {{ trans('user:login.dialog.button.signup') }}
+						router-link#register-link(to="/register") {{ trans('user:login.dialog.button.signup') }}
 </template>
 
 <script type="text/babel">
