@@ -1,13 +1,11 @@
-/* global __PASSPROTECT_CONFIG__ */
+/* global __PASSPROTECT_CONFIG__, localStorage */
 
 'use strict'
 
-import {pick} from 'lodash'
 import createSession from './createSession.gql'
 import registerUser from './registerUser.gql'
 import {generateIV, generateKey, encrypt, decrypt, createKeyDerivation} from '../../utils/crypto'
 import {parseErrors} from '../../utils/errors'
-import Promise from 'bluebird'
 
 const config = __PASSPROTECT_CONFIG__.crypto
 
@@ -15,56 +13,51 @@ export let SESSION = {
   authenticated: false
 }
 
-export function login (context, creds, redirect) {
-  return context.$apollo
-    .mutate({
+export async function login (context, creds, redirect) {
+  try {
+    const response = await context.$apollo.mutate({
       mutation: createSession,
-      variables: {
-        input: creds
-      }
+      variables: { input: creds }
     })
-    .tap(result => parseErrors(result.data.createSession))
-    .then(response => {
-      SESSION.jwtToken = response.data.createSession.token
 
-      return createClearKey(response.data.createSession.user, creds.password)
-    })
-    .then(clearKey => {
-      SESSION.username = creds.username
-      SESSION.clearKey = clearKey
-      SESSION.authenticated = true
+    parseErrors(response.data.createSession)
 
-      localStorage.setItem('username', SESSION.username)
-      localStorage.setItem('jwtToken', SESSION.jwtToken)
-      localStorage.setItem('clearKey', SESSION.clearKey)
+    const clearKey = await createClearKey(response.data.createSession.user, creds.password)
 
-      if (redirect) {
-        context.$router.go(redirect)
-      } else {
-        context.$router.push('/items')
-      }
-    })
-    .catch(err => (context.error = err))
+    SESSION.jwtToken = response.data.createSession.token
+    SESSION.username = creds.username
+    SESSION.clearKey = clearKey
+    SESSION.authenticated = true
+
+    localStorage.setItem('jwtToken', SESSION.jwtToken)
+    localStorage.setItem('username', SESSION.username)
+    localStorage.setItem('clearKey', SESSION.clearKey)
+
+    if (redirect) {
+      context.$router.go(redirect)
+    } else {
+      context.$router.push('/items')
+    }
+  } catch (err) {
+    context.error = err
+  }
 }
 
-export function signup (context, creds, redirect) {
+export async function signup (context, creds, redirect) {
   // Idem login avec mutation signup + récup token
-  return generateMasterKey(creds.username, creds.password)
-  .then(function (encryption) {
-    return context.$apollo.mutate({
+  try {
+    const encryption = await generateMasterKey(creds.username, creds.password)
+    const result = await context.$apollo.mutate({
       mutation: registerUser,
-      variables: {
-        input: {
-          _id: creds.username,
-          password: creds.password,
-          encryption
-        }
-      }
+      variables: { input: { _id: creds.username, password: creds.password, encryption } }
     })
-  })
-  .tap(result => parseErrors(result.data.registerUser))
-  .then(() => login(context, creds, redirect))
-  .catch(err => (context.error = err))
+
+    parseErrors(result.data.registerUser)
+
+    return await login(context, creds, redirect)
+  } catch (err) {
+    context.error = err
+  }
 }
 
 export function logout (context) {
@@ -94,31 +87,25 @@ export function getAuthHeader () {
   }
 }
 
-function generateMasterKey (user, password) {
-  const salt = generateIV(config.ivSize)
-  const masterKey = generateKey(config.keySize)
-  const masterKeyKey = salt.then(salt => createKeyDerivation(password, salt, config.pbkdf2))
+async function generateMasterKey (user, password) {
+  const salt = await generateIV(config.ivSize)
+  const masterKey = await generateKey(config.keySize)
+  const masterKeyKey = await createKeyDerivation(password, salt, config.pbkdf2)
 
-  const encryptedKey = Promise
-    .props({masterKey, masterKeyKey})
-    .then(obj => encrypt(obj.masterKey, obj.masterKeyKey.key, obj.masterKeyKey.iv, config.cypherIv))
+  const encryptedKey = await encrypt(masterKey, masterKeyKey.key, masterKeyKey.iv, config.cypherIv)
 
-  return Promise
-    .props({salt, encryptedKey})
-    .then(obj => pick(obj, ['salt', 'encryptedKey']))
+  return {salt, encryptedKey}
 }
 
-function createClearKey (user, password) {
+async function createClearKey (user, password) {
   const salt = user.encryption.salt
-  const masterKeyKeyPromise = createKeyDerivation(password, salt, config.pbkdf2)
+  const masterKeyKey = await createKeyDerivation(password, salt, config.pbkdf2)
 
-  return masterKeyKeyPromise
-    .then(masterKeyKey => {
-      const key = masterKeyKey.key
-      const iv = masterKeyKey.iv
-      const encryptedKey = user.encryption.encryptedKey
+  const key = masterKeyKey.key
+  const iv = masterKeyKey.iv
+  const encryptedKey = user.encryption.encryptedKey
 
-      return decrypt(encryptedKey, key, iv, config.cypherIv)
-    })
-    .then(key => key.toString('binary'))
+  const clearKey = await decrypt(encryptedKey, key, iv, config.cypherIv)
+
+  return clearKey.toString('binary')
 }
