@@ -1,46 +1,44 @@
 'use strict'
 
-import 'any-promise/register/bluebird'
+import 'babel-polyfill'
 
 // The Vue build version to load with the `import` command
 // (runtime-only or standalone) has been set in webpack.base.conf with an alias.
 import Vue from 'vue'
 import App from './App'
 import router from './router'
+import promisify from 'es6-promisify'
 
 import {init as i18nInit} from 'i18next'
-import resBundle from 'i18next-resource-store-loader!../common/locales'
-import Promise from 'bluebird'
+import resBundle from '../common/locales'
 import Vuetify from 'vuetify'
 import veeDictionary from 'vee-validate/dist/locale/fr'
 import VeeValidate, { Validator } from 'vee-validate'
 
+import VueApollo from 'vue-apollo'
 import {apolloProvider} from './utils/graphql'
 import {checkAuth} from './components/user/UserService'
 
-global.Promise = Promise
+(async function () {
+  Validator.addLocale(veeDictionary)
 
-Validator.addLocale(veeDictionary)
+  Vue.config.productionTip = false
 
-Vue.config.productionTip = false
+  Vue.use(VueApollo)
+  Vue.use(Vuetify)
+  Vue.use(VeeValidate, {errorBagName: 'veeErrors', locale: 'fr'})
 
-Vue.use(Vuetify)
-Vue.use(VeeValidate, {errorBagName: 'veeErrors', locale: 'fr'})
+  const i18nOptions = {
+    resources: resBundle,
+    lng: 'fr-FR',
+    joinArrays: '+'
+  }
 
-const i18nOptions = {
-  resources: resBundle,
-  lng: 'fr-FR',
-  joinArrays: '+'
-}
+  const trans = await promisify(i18nInit)(i18nOptions)
 
-const i18nInitPromise = Promise
-  .fromCallback((cb) => i18nInit(i18nOptions, cb))
-  .then(trans => {
-    window.trans = trans
-    Vue.prototype.trans = trans
-  })
+  window.trans = trans
+  Vue.prototype.trans = trans
 
-i18nInitPromise.then(() => {
   checkAuth()
 
   /* eslint-disable no-new */
@@ -51,4 +49,4 @@ i18nInitPromise.then(() => {
     template: '<App/>',
     components: { App }
   })
-})
+})()
